backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.3
groq==0.11.0
chromadb==1.5.5
redis[hiredis]==5.1.1
pymupdf==1.24.11
sqlalchemy==2.0.35
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
msgpack==1.1.0
brotli==1.1.0
numpy==1.26.4
scikit-learn==1.5.2
torch==2.5.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu
sentence-transformers==3.3.1
httpx==0.27.2
python-dotenv==1.0.1
pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
